Use shared Html template
